Fix default options behaviour. Now options used in constructor overrides the default ones. Also changed `resultKey` and `elementContentKey` default value.
Assign the `.complete-it` class to main element.

Change the remote query strategy.
When it is necessary to make a remote query (if it is not already
present in cache), it is triggered a custom event (`performQuery`)
to the main `$element`.
In this way the related callback `performQuery` can be easily
throttled.

Perform the exact match between query input and results by
lowercasing results elements.
Fix the `indexOf` condition.
Remove the `id` from each results to reduce memory usage.

Fix the sort order (It is now descending).