Add cache strategy and `mapAndFilter` option.
It is now possible, using the option `cache`, to set a cache
strategy. Allowed values are `false`, `memory`, `sessionStorage`
and `localStorage`.

The introduction of `mapAndFilter` option is a consequence of a
new design choice. It is assumed that the result of the ajax query
is an array of autocomplete objects, ordered and with a standard
form.
The ajax response it is passed to `mapAndFilter` to be altered and
standardized to the described form.